tools/jira-component-mapper/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==1.7.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
defusedxml==0.7.1
Deprecated==1.2.13
docker-pycreds==0.4.0
docopt==0.6.2
et-xmlfile==1.1.0
Flask==2.2.2
flask-marshmallow==0.14.0
frozenlist==1.3.3
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jira==3.5.2
MarkupSafe==2.1.1
marshmallow==3.19.0
multidict==6.0.4
neo4j==5.2.1
numpy==1.22.4
oauthlib==3.2.2
openai==0.19.0
openpyxl==3.0.10
packaging==21.3
pandas==1.4.2
pandas-stubs==1.2.0.61
path==16.5.0
pathtools==0.1.2
pipreqs==0.4.11
promise==2.3
protobuf==3.20.1
psutil==5.9.1
pycodestyle==2.9.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.1
PyYAML==6.0
redis==4.3.4
requests==2.28.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
sentry-sdk==1.5.12
setproctitle==1.2.3
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
toml==0.10.2
tqdm==4.64.0
typing_extensions==4.7.1
urllib3==1.26.9
wandb==0.12.18
Werkzeug==2.2.2
wrapt==1.14.1
yarg==0.1.9
yarl==1.8.2
zipp==3.10.0
